Move client logic to client.rs
